I am posting the initial versions of the MATLAB Code and spreadsheet to this branch
I haven't made changes to these files, they are the initial versions created by me, the MATLAB code should be altered to reflect the GNU v3 license in the repo before it is merged. This will be done in the following commit to spreadsheet-updates, just wanted to get the current versions up.